8260264: Move common os_<unix> inline methods to a common posix source file
