+++ src/hotspot/os/aix/os_aix.inline.hpp
@@ -32,14 +32,6 @@
 // System includes
 
 #include <unistd.h>
-#include <sys/socket.h>
-#include <poll.h>
-#include <sys/ioctl.h>
-#include <netdb.h>
-
-inline bool os::uses_stack_guard_pages() {
-  return true;
-}
 
 // Whether or not calling code should/can commit/uncommit stack pages
 // before guarding them. Answer for AIX is definitly no, because memory
@@ -49,22 +41,6 @@ inline bool os::must_commit_stack_guard_pages() {
   return false;
 }
 
-// Bang the shadow pages if they need to be touched to be mapped.
-inline void os::map_stack_shadow_pages(address sp) {
-}
-
-inline void os::dll_unload(void *lib) {
-  ::dlclose(lib);
-}
-
-inline jlong os::lseek(int fd, jlong offset, int whence) {
-  return (jlong) ::lseek64(fd, offset, whence);
-}
-
-inline int os::fsync(int fd) {
-  return ::fsync(fd);
-}
-
 inline int os::ftruncate(int fd, jlong length) {
   return ::ftruncate64(fd, length);
 }
@@ -72,43 +48,5 @@ inline int os::ftruncate(int fd, jlong length) {
 // We don't have NUMA support on Aix, but we need this for compilation.
 inline bool os::numa_has_static_binding()   { ShouldNotReachHere(); return true; }
 inline bool os::numa_has_group_homing()     { ShouldNotReachHere(); return false;  }
-
-inline size_t os::write(int fd, const void *buf, unsigned int nBytes) {
-  size_t res;
-  RESTARTABLE((size_t) ::write(fd, buf, (size_t) nBytes), res);
-  return res;
-}
-
-inline int os::socket_close(int fd) {
-  return ::close(fd);
-}
-
-inline int os::socket(int domain, int type, int protocol) {
-  return ::socket(domain, type, protocol);
-}
-
-inline int os::recv(int fd, char* buf, size_t nBytes, uint flags) {
-  RESTARTABLE_RETURN_INT(::recv(fd, buf, nBytes, flags));
-}
-
-inline int os::send(int fd, char* buf, size_t nBytes, uint flags) {
-  RESTARTABLE_RETURN_INT(::send(fd, buf, nBytes, flags));
-}
-
-inline int os::raw_send(int fd, char *buf, size_t nBytes, uint flags) {
-  return os::send(fd, buf, nBytes, flags);
-}
-
-inline int os::connect(int fd, struct sockaddr *him, socklen_t len) {
-  RESTARTABLE_RETURN_INT(::connect(fd, him, len));
-}
-
-inline struct hostent* os::get_host_by_name(char* name) {
-  return ::gethostbyname(name);
-}
-
-inline void os::exit(int num) {
-  ::exit(num);
-}
 
 #endif // OS_AIX_OS_AIX_INLINE_HPP
